Fixed compilation.
